requirements.txt
h11~=0.14.0
pip~=25.0
solders~=0.23.0
solana~=0.36.3
construct~=2.10.68
sniffio~=1.3.1
httpcore~=1.0.7
idna~=3.10
httpx~=0.28.1
psutil~=6.1.1
typing_extensions~=4.12.2
websockets~=13.1
certifi~=2025.1.31
anyio~=4.8.0
python-telegram-bot~=21.10